+++ test/testdata/misc_py39.py
@@ -10,6 +10,9 @@ from typing import NamedTuple
 
 
 class NamedTupleExample(NamedTuple):
+    """
+    An example for a typing.NamedTuple.
+    """
     name: str
     """Name of our example tuple."""
     id: int = 3